+++ src/Powercord/plugins/pc-emojiUtility/manifest.json
@@ -1,0 +1,11 @@
+{
+  "name": "Emote Utility",
+  "version": "1.0.0",
+  "description": "Emote related commands",
+  "author": "Joakim#9814",
+  "license": "MIT",
+  "repo": null,
+  "dependencies": [
+    "pc-commands"
+  ]
+}
